% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIMA-package.R
\docType{package}
\name{CHIMA}
\alias{CHIMA}
\alias{CHIMA-package}
\title{CHIMA: The CHIMA package provides tools for high-dimensional mediator screening and testing}
\description{
The CHIMA package provides tools for high-dimensional mediator screening and testing using Ridge-HOLP and Approximate Orthogonalization methods.
}
\section{Vignettes}{

See \code{vignette("getting-started", package = "CHIMA")} for a walkthrough.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/samuelOsarfo/CHIMA}
  \item Report bugs at \url{https://github.com/samuelOsarfo/CHIMA/issues}
}

}
\author{
\strong{Maintainer}: Sangyoon Yi \email{sayi@okstate.edu}

Authors:
\itemize{
  \item Samuel Osarfo \email{sosarfo@okstate.edu}
}

}
\keyword{Analysis}
\keyword{Brief}
\keyword{CHIMA:}
\keyword{Dimensional}
\keyword{High}
\keyword{Mediation}
\keyword{Package}
\keyword{does.}
\keyword{for}
\keyword{internal}
\keyword{of}
\keyword{overview}
\keyword{package}
\keyword{the}
\keyword{what}
